In this exercise, one can expect a list of samples on Iris flowers. Each sample is represented with a tuple of two values: <petal_length, species>, where the first attribute is the measurement on the length of the petal for the sample, and the second attribute indicates the species of sample. Here is an example.
The task is to split the sample list into two sublists, while complying with the following two conditions:

The petal_length of sample in one sublist is less or equal than that of any sample in the other sublist.
The information gain on the species attribute of the sublists is maximal among all possible splits.
As output, one should just return the information gain.

In addition, one can expect that each value of petal_length is unique.